Downgrade python to 3.12 (#640)
* Downgrade python to 3.12

We are now able to install node's dependencies on the fly,
but the python version needs to be the same as the engine.
The sentencepiece depedency seems to currently only work
for python 3.12, so I'll need to downgrade the engine to
3.12 until either a workaround is found for 3.13 or we remove
the need for the python versions to be the same.

* missed some places

* fix check/format no exclude in 3.12

* use typing_extensions

* format