feat: add fork/merge/sync operations (#1750)
* feat: add fork/merge/sync operations

This PR adds a few key features:

This manager is responsible for syncing to/from a sync backend. Right
now this is only GTC but it could easily be generalized to other
backends (S3).

After engine initialization, the sync manager pulls down `.py` files
from the bucket configured in `GT_CLOUD_BUCKET_ID`. The manager also
sets up file watchers (using watchdog) for create/update/delete
operations in the `synced_workflows` directory.

This request accepts an existing workflow and creates a fork of it. This
is effectively a copy, but with some metadata in the fork about where it
came from.

Takes a workflow fork and merges it into its source.

Takes a workflow fork and replaces its contents with its source.

Accepts two workflows and compares them to see who is newer. It does
this by comparing metadata (`last_modified_date`).

This request has been slightly changed when saving workflows in a synced
directory. Similar to when saving templates, saving a workflow from a
synced directory will automatically create a fork (with a max of 1).

1. User A creates a workflow.
2. User A adds this workflow to the synced_workflows directory.
3. Workflow is automatically synced up to GTC bucket.
4  User B starts their engine.
5. User B has the workflow placed in their synced_workflows directory.
6. User B starts working on this, saves the file.
7. User B receives a fork of the workflow.
8. User B continues to work on the workflow.
9. User B is satisfied with the workflow, selects "Merge Fork Into
   Source".
10. Workflow is automatically synced up to GTC bucket.

* fixup! feat: add fork/merge/sync operations

* fixup! feat: add fork/merge/sync operations