refactor: remove unneeded active node check (#1974)
Getting this error when starting a flow, cancelling, and then starting
again.

I suspect the flow is not being fully cancelled so run 1's
`NodeFinishProcessEvent` comes across as run 2 is starting.

Will create a followup task to investigate proper cancellation, but I
don't see much value in having this error.