refactor: move static and mcp server initialization into managers
The static and mcp servers are tightly coupled to the lifecycle of
app.py. i.e. they don't start without running `gtn`. This means that
executors/tests/etc _using_ the engine need to manually spin up the
servers.

This is problematic and difficult to maintain.

This PR moves the servers into their respective manager classes. This
also let's us use more information when starting up the servers such as
the storage backend when starting the static server.