fix: remove broken flush in queue check (#3145)
Somehow, sometimes, we enter a deadlock where is wrongly
self._flush_in_queue=True. This prevents all future updates from being
emitted. This is effectively an thread unsafe lock and has likely become
a more prevelant issue as we've asyncified things.

Removing, does not appear to be a required optimization.